medical-chatbot-backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# OpenAI
openai==1.10.0

# HTTP requests for tools
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Utilities
python-dateutil==2.8.2
